+++ library/core/src/main/java/com/google/android/exoplayer2/source/chunk/SingleSampleMediaChunk.java
@@ -34,7 +34,6 @@ public final class SingleSampleMediaChunk extends BaseMediaChunk {
   private final Format sampleFormat;
 
   private volatile int bytesLoaded;
-  private volatile boolean loadCanceled;
   private volatile boolean loadCompleted;
 
   /**
@@ -45,7 +44,7 @@ public final class SingleSampleMediaChunk extends BaseMediaChunk {
    * @param trackSelectionData See {@link #trackSelectionData}.
    * @param startTimeUs The start time of the media contained by the chunk, in microseconds.
    * @param endTimeUs The end time of the media contained by the chunk, in microseconds.
-   * @param chunkIndex The index of the chunk.
+   * @param chunkIndex The index of the chunk, or {@link C#INDEX_UNSET} if it is not known.
    * @param trackType The type of the chunk. Typically one of the {@link C} {@code TRACK_TYPE_*}
    *     constants.
    * @param sampleFormat The {@link Format} of the sample in the chunk.
@@ -90,12 +89,7 @@ public final class SingleSampleMediaChunk extends BaseMediaChunk {
 
   @Override
   public void cancelLoad() {
-    loadCanceled = true;
-  }
-
-  @Override
-  public boolean isLoadCanceled() {
-    return loadCanceled;
+    // Do nothing.
   }
 
   @SuppressWarnings("NonAtomicVolatileUpdate")