+++ library/core/src/main/java/com/google/android/exoplayer2/trackselection/MappingTrackSelector.java
@@ -16,6 +16,7 @@
 package com.google.android.exoplayer2.trackselection;
 
 import android.support.annotation.IntDef;
+import android.support.annotation.Nullable;
 import android.util.Pair;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.ExoPlaybackException;
@@ -301,13 +302,13 @@ public abstract class MappingTrackSelector extends TrackSelector {
 
   }
 
-  private MappedTrackInfo currentMappedTrackInfo;
+  private @Nullable MappedTrackInfo currentMappedTrackInfo;
 
   /**
    * Returns the mapping information for the currently active track selection, or null if no
    * selection is currently active.
    */
-  public final MappedTrackInfo getCurrentMappedTrackInfo() {
+  public final @Nullable MappedTrackInfo getCurrentMappedTrackInfo() {
     return currentMappedTrackInfo;
   }
 
@@ -357,9 +358,11 @@ public abstract class MappingTrackSelector extends TrackSelector {
     int[] rendererTrackTypes = new int[rendererCapabilities.length];
     for (int i = 0; i < rendererCapabilities.length; i++) {
       int rendererTrackGroupCount = rendererTrackGroupCounts[i];
-      rendererTrackGroupArrays[i] = new TrackGroupArray(
-          Arrays.copyOf(rendererTrackGroups[i], rendererTrackGroupCount));
-      rendererFormatSupports[i] = Arrays.copyOf(rendererFormatSupports[i], rendererTrackGroupCount);
+      rendererTrackGroupArrays[i] =
+          new TrackGroupArray(
+              Util.nullSafeArrayCopy(rendererTrackGroups[i], rendererTrackGroupCount));
+      rendererFormatSupports[i] =
+          Util.nullSafeArrayCopy(rendererFormatSupports[i], rendererTrackGroupCount);
       rendererTrackTypes[i] = rendererCapabilities[i].getTrackType();
     }
 
@@ -367,7 +370,7 @@ public abstract class MappingTrackSelector extends TrackSelector {
     int unmappedTrackGroupCount = rendererTrackGroupCounts[rendererCapabilities.length];
     TrackGroupArray unmappedTrackGroupArray =
         new TrackGroupArray(
-            Arrays.copyOf(
+            Util.nullSafeArrayCopy(
                 rendererTrackGroups[rendererCapabilities.length], unmappedTrackGroupCount));
 
     // Package up the track information and selections.
@@ -399,11 +402,12 @@ public abstract class MappingTrackSelector extends TrackSelector {
    *     RendererCapabilities#getTrackType()} is {@link C#TRACK_TYPE_NONE}.
    * @throws ExoPlaybackException If an error occurs while selecting the tracks.
    */
-  protected abstract Pair<RendererConfiguration[], TrackSelection[]> selectTracks(
-      MappedTrackInfo mappedTrackInfo,
-      int[][][] rendererFormatSupports,
-      int[] rendererMixedMimeTypeAdaptationSupport)
-      throws ExoPlaybackException;
+  protected abstract Pair<RendererConfiguration[], TrackSelection[]>
+      selectTracks(
+          MappedTrackInfo mappedTrackInfo,
+          int[][][] rendererFormatSupports,
+          int[] rendererMixedMimeTypeAdaptationSupport)
+          throws ExoPlaybackException;
 
   /**
    * Finds the renderer to which the provided {@link TrackGroup} should be mapped.